refactor: 상품상세페이지 parallel route 사용해보기
