+++ src/app/(common)/(market)/items/[id]/@comments/loading.tsx
@@ -1,0 +1,5 @@
+import { Message } from "@/components/ui";
+
+export default function Loading() {
+  return <Message>코멘트를 가져오는중입니다...</Message>;
+}
