+++ src/app/(common)/(market)/items/[id]/@detail/loading.tsx
@@ -1,0 +1,5 @@
+import { Message } from "@/components/ui";
+
+export default function Loading() {
+  return <Message>상품상제정보를 가져오는중입니다...</Message>;
+}
